pyproject.toml
[project]
name = "ip-dnb"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tinydb>=4.8.2",
    "typer>=0.24.0",
]

[project.scripts]
ip-dnb = "ip_dnb.main:app"


[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
